chore: Remove Prettier in favor of `forge fmt`

+++ src/UniswapV2Router02/contracts/interfaces/IWETH.sol
@@ -2,9 +2,9 @@ pragma solidity =0.6.6;
 pragma experimental ABIEncoderV2;
 
 interface IWETH {
-	function deposit() external payable;
+    function deposit() external payable;
 
-	function transfer(address to, uint value) external returns (bool);
+    function transfer(address to, uint256 value) external returns (bool);
 
-	function withdraw(uint) external;
+    function withdraw(uint256) external;
 }
